+++ factory/utils.py
@@ -1,11 +1,8 @@
-# -*- coding: utf-8 -*-
 # Copyright: See the LICENSE file.
 
-from __future__ import unicode_literals
 
 import collections
-
-from . import compat
+import importlib
 
 
 def import_object(module_name, attribute_name):
@@ -15,19 +12,11 @@ def import_object(module_name, attribute_name):
         >>> import_object('datetime', 'datetime')
         <type 'datetime.datetime'>
     """
-    # Py2 compatibility: force str (i.e bytes) when importing.
-    module = __import__(str(module_name), {}, {}, [str(attribute_name)], 0)
-    return getattr(module, str(attribute_name))
+    module = importlib.import_module(module_name)
+    return getattr(module, attribute_name)
 
 
-def _safe_repr(obj):
-    try:
-        return log_repr(obj)
-    except Exception:
-        return '<bad_repr object at %s>' % id(obj)
-
-
-class log_pprint(object):
+class log_pprint:
     """Helper for properly printing args / kwargs passed to an object.
 
     Since it is only used with factory.debug(), the computation is
@@ -45,26 +34,18 @@ class log_pprint(object):
     def __str__(self):
         return ', '.join(
             [
-                _safe_repr(arg) for arg in self.args
+                repr(arg) for arg in self.args
             ] + [
-                '%s=%s' % (key, _safe_repr(value))
+                '%s=%s' % (key, repr(value))
                 for key, value in self.kwargs.items()
             ]
         )
 
 
-def log_repr(obj):
-    """Generate a text-compatible repr of an object.
-
-    Some projects have a tendency to generate bytes-style repr in Python2.
-    """
-    return compat.force_text(repr(obj))
-
-
-class ResetableIterator(object):
+class ResetableIterator:
     """An iterator wrapper that can be 'reset()' to its start."""
     def __init__(self, iterator, **kwargs):
-        super(ResetableIterator, self).__init__(**kwargs)
+        super().__init__(**kwargs)
         self.iterator = iter(iterator)
         self.past_elements = collections.deque()
         self.next_elements = collections.deque()
@@ -87,7 +68,7 @@ class ResetableIterator(object):
         self.next_elements.extend(self.past_elements)
 
 
-class OrderedBase(object):
+class OrderedBase:
     """Marks a class as being ordered.
 
     Each instance (even from subclasses) will share a global creation counter.
@@ -96,7 +77,7 @@ class OrderedBase(object):
     CREATION_COUNTER_FIELD = '_creation_counter'
 
     def __init__(self, **kwargs):
-        super(OrderedBase, self).__init__(**kwargs)
+        super().__init__(**kwargs)
         if type(self) is not OrderedBase:
             self.touch_creation_counter()