+++ tests/djapp/settings.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Copyright: See the LICENSE file.
 
 """Settings for factory_boy/Django tests."""